PHP 8.5: Remove deprecated non-canonical scalar type casts
Summary:
PHP 8.5 deprecates non-canonical scalar type casts.
See https://wiki.php.net/rfc/deprecations_php_8_5#deprecate_non-standard_cast_names and https://wiki.php.net/rfc/deprecations_php_8_5

Closes T16317

Test Plan: In theory, run `./bin/arc unit --everything` (which fails with at least three known issues for me, sigh)

Reviewers: O1 Blessed Committers, mainframe98

Reviewed By: O1 Blessed Committers, mainframe98

Subscribers: tobiaswiese, valerio.bozzolan, Matthew, Cigaryno

Maniphest Tasks: T16317

Differential Revision: https://we.phorge.it/D26444

+++ src/applications/conduit/parametertype/ConduitPointsParameterType.php
@@ -14,7 +14,7 @@ final class ConduitPointsParameterType
     }
 
     if ($value !== null) {
-      $value = (double)$value;
+      $value = (float)$value;
       if ($value < 0) {
         $this->raiseValidationException(
           $request,
