+++ src/applications/search/management/PhabricatorSearchManagementNgramsWorkflow.php
@@ -48,7 +48,7 @@ final class PhabricatorSearchManagementNgramsWorkflow
           pht('Specify a numeric threshold between 0 and 1.'));
       }
 
-      $threshold = (double)$threshold;
+      $threshold = (float)$threshold;
       if ($threshold <= 0 || $threshold >= 1) {
         throw new PhutilArgumentUsageException(
           pht('Threshold must be greater than 0.0 and less than 1.0.'));
